pyproject.toml
[tool.poetry]
name = "capstone-recpiebox"
version = "0.1.0"
description = ""
authors = ["cbw29512 <divclass01@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.8"
psycopg2-binary = "^2.8.5"
django-allauth = "^0.42.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
